requirements.txt
Flask==2.2.3
flask-cors==3.0.10

Werkzeug<2.3.0

Pillow


# Core ML & data libraries
torch>=1.13.0
torchvision>=0.14.0

# Data handling & progress bars
pandas>=1.5.0
tqdm>=4.64.0
numpy>=1.23.0

# Image I/O & preprocessing
opencv-python>=4.6.0
Pillow>=9.4.0

# Plotting
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine learning utilities
scikit-learn>=1.1.0

# Experiment tracking
wandb>=0.13.0

